requirements.txt
django>=5.2.3
djangorestframework
djangorestframework-simplejwt
gunicorn
